(BSR)[PRO] refactor: remove accessibility from serializer
