+++ pro/src/core/Venue/adapters/__specs__/useGetVenue.spec.ts
@@ -2,7 +2,6 @@ import { renderHook, waitFor } from '@testing-library/react'
 
 import { api } from 'apiClient/api'
 import { GetVenueResponseModel, VenueTypeCode } from 'apiClient/v1'
-import { AccessiblityEnum } from 'core/shared'
 
 import { useGetVenue } from '../getVenueAdapter'
 
@@ -80,13 +79,6 @@ describe('useGetVenue', () => {
     expect(loadingState.error).toBeUndefined()
 
     const venue = {
-      accessibility: {
-        [AccessiblityEnum.AUDIO]: false,
-        [AccessiblityEnum.MENTAL]: false,
-        [AccessiblityEnum.MOTOR]: false,
-        [AccessiblityEnum.VISUAL]: false,
-        [AccessiblityEnum.NONE]: true,
-      },
       address: '12 rue du pilas',
       bannerMeta: {
         image_credit: '',
