+++ pro/src/core/Venue/adapters/getVenueAdapter/serializers.ts
@@ -1,27 +1,12 @@
 import { BannerMetaModel, GetVenueResponseModel } from 'apiClient/v1'
 import { VenueBannerMetaProps } from 'components/VenueForm/ImageUploaderVenue/ImageUploaderVenue'
-import { AccessiblityEnum } from 'core/shared'
 import { Venue } from 'core/Venue/types'
 import { getLastCollectiveDmsApplication } from 'utils/getLastCollectiveDmsApplication'
 
 export const serializeVenueApi = (venue: GetVenueResponseModel): Venue => {
-  const venueAccessibility = {
-    [AccessiblityEnum.VISUAL]: venue.visualDisabilityCompliant || false,
-    [AccessiblityEnum.MENTAL]: venue.mentalDisabilityCompliant || false,
-    [AccessiblityEnum.AUDIO]: venue.audioDisabilityCompliant || false,
-    [AccessiblityEnum.MOTOR]: venue.motorDisabilityCompliant || false,
-    [AccessiblityEnum.NONE]: [
-      venue.visualDisabilityCompliant,
-      venue.mentalDisabilityCompliant,
-      venue.audioDisabilityCompliant,
-      venue.motorDisabilityCompliant,
-    ].every((accessibility) => accessibility === false),
-  }
-
   /* istanbul ignore next: DEBT, TO FIX */
   return {
     ...venue,
-    accessibility: venueAccessibility,
     bannerMeta: venue.bannerMeta
       ? serializeBannerMetaApi(venue.bannerMeta)
       : null,
